fix(test): SMI-4754 async createDatabase in EnterpriseAuditLogger tests
Replace deprecated sync createDatabase() with createDatabaseAsync() + initializeSchema()
across all 11 call-sites in the enterprise AuditLogger test suite. Post-WASM migration
(SMI-2721, Feb 2026), the sync factory was unreliable in the virtiofs worktree
environment — db initialisation would not complete before the EnterpriseAuditLogger
constructor ran, leaving logger undefined in afterEach and causing 43/43 tests to fail:

  TypeError: Cannot read properties of undefined (reading 'dispose')
    at tests/AuditLogger.test.ts:50:12

Split original 857-line file into two companion files to satisfy the 500-line gate
(audit:standards Check):
  - AuditLogger.test.ts      (426 lines): constructor, exporter registration,
                                           logSSOEvent, logRBACEvent, logLicenseEvent
  - AuditLogger.flush.test.ts (493 lines): flush, auto-flush, queryEnterprise, dispose,
                                           getBufferSize, retention policy, event type
                                           mapping, integration tests

No production code changes. RetentionEnforcer.test.ts was already on createDatabaseAsync.

Before: 43 failing / 43 total (0 passing); full suite 524/567 (43 broken)
After:  43 passing / 43 total; full suite 567/567
